Require ExcludeAssets=runtime on refs
PrivateAssets controls the flow of dependencies between NuGet packages,
but not whether a given project copies package assets to its output
folder.

ExcludeAssets=runtime indicates that the runtime assets provided
(transitively, evidently) by the package should not be considered as
project outputs, which is what we want for this package: don't copy
MSBuild assemblies to your output.