Update step: Ensure place for initialization
The .NET Runtime loads assemblies required to execute a method just before entering the method.

That means that MSLocator must run in a separate method, before any method that uses the Microsoft.Build namespace.

This commit creates space for that by extracting the Build() method. In a larger real project, this may already be done.