Add explanation for Builder class in example
When implementing MSBuildLocator I originally called `MSBuildLocator.RegisterInstance()`, then later in the same function I called other Microsoft.Build functions. 

This is my first large project with a major C# component. I didn't know that all of the dll imports in my function would be loaded when the function was entered, instead of when the import procedures were called.

This request adds a short description above the Builder class to explain why it's necessary--so that the MSBuildLocator dll loads first, then after the appropriate setup is requested and the `Builder` class is instantiated can more Microsoft.Build dlls be loaded properly.