Disable GitVersioning "CloudBuild" versioning
VSTS public builds currently disallow specifying the version from within
the build job, so disable it for PR jobs. This will not affect the version
of the produced binaries, only the build number specified in the job.

Instead, set the number from a single (new) project, and build it
conditionally on not-PR.