Use hostfxr hosting layer API (#161)
This makes the process of finding .NET SDK instances more efficient and fixes dotnet/msbuild#7466

This uses hostfxr, similar to what the MSBuildSdkResolver does in dotnet/sdk, to find sdks.

Couple important notes:
realpath is for following symlinks to real dotnet instances. On linux, it is common for PATH to include a symlink to the folder that actually has dotnet rather than the folder itself.
This throws exceptions if either native method call fails.